+++ Realm/Realm/Exceptions/RealmExceptionCodes.cs
@@ -36,6 +36,8 @@ namespace Realms.Exceptions
         RealmInvalidTransaction = 10,
         RealmFormatUpgradeRequired = 13,
         RealmSchemaMismatch = 14,
+        RealmSchemaValidation = 15,
+
         RealmRowDetached = 21,
         RealmTableHasNoPrimaryKey = 22,
         RealmDuplicatePrimaryKeyValue = 23,