+++ Realm/Realm/Handles/SetHandle.cs
@@ -125,7 +125,7 @@ namespace Realms
             return new ThreadSafeReferenceHandle(result);
         }
 
-        public override ResultsHandle GetFilteredResults(string query)
+        public override ResultsHandle GetFilteredResults(string query, RealmValue[] arguments)
         {
             throw new NotImplementedException("Sets can't be filtered yet.");
         }