+++ Tests/Benchmarks/Benchmarks.iOS/AppDelegate.cs
@@ -1,0 +1,43 @@
+﻿////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+using System.Linq;
+using Foundation;
+using UIKit;
+using Xamarin.Forms;
+using Xamarin.Forms.Platform.iOS;
+
+namespace Benchmarks.iOS
+{
+    [Register("AppDelegate")]
+    public partial class AppDelegate : FormsApplicationDelegate
+    {
+        public override bool FinishedLaunching(UIApplication app, NSDictionary options)
+        {
+            var arguments = NSProcessInfo.ProcessInfo.Arguments
+                .Select(a => a.Replace("-app-arg=", string.Empty))
+                .Skip(1).ToArray();
+
+            Forms.Init();
+            LoadApplication(new App(arguments));
+
+            return base.FinishedLaunching(app, options);
+
+        }
+    }
+}